backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pdfplumber
python-docx
google-generativeai
python-multipart
